# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A movie management REST API. Movie endpoints require authentication via a Bearer token obtained from POST /v1/tokens/authentication.",
        "title": "Greenlight Restful API",
        "contact": {},
        "version": "1.1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/v1/tokens/authentication": {
            "post": {
                "description": "Validates email and password credentials and returns a 24-hour Bearer token to use in the Authorization header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tokens"
                ],
                "summary": "Create an authentication token",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.createAuthTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.authTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.errorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/main.validationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.authTokenObject": {
            "type": "object",
            "properties": {
                "expiry": {
                    "type": "string",
                    "example": "2024-01-02T00:00:00Z"
                },
                "token": {
                    "type": "string",
                    "example": "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
                }
            }
        },
        "main.authTokenResponse": {
            "type": "object",
            "properties": {
                "authentication_token": {
                    "$ref": "#/definitions/main.authTokenObject"
                }
            }
        },
        "main.createAuthTokenRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "alice@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
...
        "main.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "the requested resource could not be found"
                }
            }
        },
...
        "main.validationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the \"Bearer \" prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}